# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type CareCompassRequestMediaType = typeof CareCompassRequestMediaType[keyof typeof CareCompassRequestMediaType];


export const CareCompassRequestMediaType = {
  'application/pdf': 'application/pdf',
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/gif': 'image/gif',
  'image/webp': 'image/webp',
} as const;

export interface CareCompassRequest {
  /**
     * @minLength 3
     * @maxLength 3000
     */
  situation: string;
  /** @maxLength 12000 */
  documentText?: string;
  /** @maxLength 255 */
  fileName?: string;
  mediaType?: CareCompassRequestMediaType;
  /** @maxLength 9800000 */
  documentBase64?: string;
}

export type CareCompassDepartmentUrgency = typeof CareCompassDepartmentUrgency[keyof typeof CareCompassDepartmentUrgency];


export const CareCompassDepartmentUrgency = {
  routine: 'routine',
  soon: 'soon',
  'urgent-seek-care-now': 'urgent-seek-care-now',
} as const;

export interface CareCompassDepartment {
  name: string;
  reason: string;
  urgency: CareCompassDepartmentUrgency;
}

export interface CareCompassMedicine {
  name: string;
  /** @nullable */
  dosage: string | null;
  /** @nullable */
  instructions: string | null;
}

export interface CareCompassExtractedInformation {
  medicines: CareCompassMedicine[];
  dates: string[];
  doctors: string[];
  previousConditions: string[];
  otherDetails: string[];
}

export interface CareCompassChecklistItem {
  id: string;
  label: string;
  checked: boolean;
}

export interface CareCompassEmergency {
  detected: boolean;
  /** @nullable */
  message: string | null;
}

export interface CareCompassPlan {
  department: CareCompassDepartment;
  extractedInformation: CareCompassExtractedInformation;
  visitChecklist: CareCompassChecklistItem[];
  /** @nullable */
  documentSummary: string | null;
  /**
     * @minItems 5
     * @maxItems 7
     */
  questionsToAskDoctor: string[];
  emergency: CareCompassEmergency;
}