# lib/api-zod/src/generated/types/careCompassChecklistItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareCompassChecklistItem {
  id: string;
  label: string;
  checked: boolean;
}