# lib/api-zod/src/generated/types/careCompassDepartment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareCompassDepartmentUrgency } from './careCompassDepartmentUrgency';

export interface CareCompassDepartment {
  name: string;
  reason: string;
  urgency: CareCompassDepartmentUrgency;
}