# lib/api-zod/src/generated/types/careCompassEmergency.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareCompassEmergency {
  detected: boolean;
  /** @nullable */
  message: string | null;
}