# lib/api-zod/src/generated/types/careCompassExtractedInformation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareCompassMedicine } from './careCompassMedicine';

export interface CareCompassExtractedInformation {
  medicines: CareCompassMedicine[];
  dates: string[];
  doctors: string[];
  previousConditions: string[];
  otherDetails: string[];
}