# lib/api-zod/src/generated/types/careCompassMedicine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareCompassMedicine {
  name: string;
  /** @nullable */
  dosage: string | null;
  /** @nullable */
  instructions: string | null;
}