# lib/api-zod/src/generated/types/careCompassPlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareCompassChecklistItem } from './careCompassChecklistItem';
import type { CareCompassDepartment } from './careCompassDepartment';
import type { CareCompassEmergency } from './careCompassEmergency';
import type { CareCompassExtractedInformation } from './careCompassExtractedInformation';

export interface CareCompassPlan {
  department: CareCompassDepartment;
  extractedInformation: CareCompassExtractedInformation;
  visitChecklist: CareCompassChecklistItem[];
  /** @nullable */
  documentSummary: string | null;
  /**
     * @minItems 5
     * @maxItems 7
     */
  questionsToAskDoctor: string[];
  emergency: CareCompassEmergency;
}