# lib/api-zod/src/generated/types/careCompassRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareCompassRequestMediaType } from './careCompassRequestMediaType';

export interface CareCompassRequest {
  /**
     * @minLength 3
     * @maxLength 3000
     */
  situation: string;
  /** @maxLength 12000 */
  documentText?: string;
  /** @maxLength 255 */
  fileName?: string;
  mediaType?: CareCompassRequestMediaType;
  /** @maxLength 9800000 */
  documentBase64?: string;
}